+++ Sources/ParseSwift/Objects/ParseObject.swift
@@ -61,7 +61,7 @@ public extension Sequence where Element: ParseObject {
 
     /**
      Saves a collection of objects *synchronously* all at once and throws an error if necessary.
-     - parameter batchLimit: The maximum number of objects to send in each batch. If the items to be batched
+     - parameter batchLimit: The maximum number of objects to send in each batch. If the items to be batched.
      is greater than the `batchLimit`, the objects will be sent to the server in waves up to the `batchLimit`.
      Defaults to 50.
      - parameter transaction: Treat as an all-or-nothing operation. If some operation failure occurs that
@@ -126,7 +126,7 @@ public extension Sequence where Element: ParseObject {
         if transaction {
             batchLimit = commands.count
         } else {
-            batchLimit = limit != nil ? limit! : ParseConstants.batchLimit
+            batchLimit = limit ?? ParseConstants.batchLimit
         }
         let batches = BatchUtils.splitArray(commands, valuesPerSegment: batchLimit)
         try batches.forEach {
@@ -143,7 +143,7 @@ public extension Sequence where Element: ParseObject {
 
     /**
      Saves a collection of objects all at once *asynchronously* and executes the completion block when done.
-     - parameter batchLimit: The maximum number of objects to send in each batch. If the items to be batched
+     - parameter batchLimit: The maximum number of objects to send in each batch. If the items to be batched.
      is greater than the `batchLimit`, the objects will be sent to the server in waves up to the `batchLimit`.
      Defaults to 50.
      - parameter transaction: Treat as an all-or-nothing operation. If some operation failure occurs that
@@ -163,8 +163,12 @@ public extension Sequence where Element: ParseObject {
         callbackQueue: DispatchQueue = .main,
         completion: @escaping (Result<[(Result<Element, ParseError>)], ParseError>) -> Void
     ) {
-        let queue = DispatchQueue(label: "com.parse.saveAll", qos: .default,
-                                  attributes: .concurrent, autoreleaseFrequency: .inherit, target: nil)
+        let uuid = UUID()
+        let queue = DispatchQueue(label: "com.parse.saveAll.\(uuid)",
+                                  qos: .default,
+                                  attributes: .concurrent,
+                                  autoreleaseFrequency: .inherit,
+                                  target: nil)
         queue.sync {
 
             var childObjects = [String: PointerType]()
@@ -220,7 +224,7 @@ public extension Sequence where Element: ParseObject {
                 if transaction {
                     batchLimit = commands.count
                 } else {
-                    batchLimit = limit != nil ? limit! : ParseConstants.batchLimit
+                    batchLimit = limit ?? ParseConstants.batchLimit
                 }
                 let batches = BatchUtils.splitArray(commands, valuesPerSegment: batchLimit)
                 var completed = 0
@@ -361,7 +365,7 @@ public extension Sequence where Element: ParseObject {
 
     /**
      Deletes a collection of objects *synchronously* all at once and throws an error if necessary.
-     - parameter batchLimit: The maximum number of objects to send in each batch. If the items to be batched
+     - parameter batchLimit: The maximum number of objects to send in each batch. If the items to be batched.
      is greater than the `batchLimit`, the objects will be sent to the server in waves up to the `batchLimit`.
      Defaults to 50.
      - parameter transaction: Treat as an all-or-nothing operation. If some operation failure occurs that
@@ -390,7 +394,7 @@ public extension Sequence where Element: ParseObject {
         if transaction {
             batchLimit = commands.count
         } else {
-            batchLimit = limit != nil ? limit! : ParseConstants.batchLimit
+            batchLimit = limit ?? ParseConstants.batchLimit
         }
         let batches = BatchUtils.splitArray(commands, valuesPerSegment: batchLimit)
         try batches.forEach {
@@ -404,7 +408,7 @@ public extension Sequence where Element: ParseObject {
 
     /**
      Deletes a collection of objects all at once *asynchronously* and executes the completion block when done.
-     - parameter batchLimit: The maximum number of objects to send in each batch. If the items to be batched
+     - parameter batchLimit: The maximum number of objects to send in each batch. If the items to be batched.
      is greater than the `batchLimit`, the objects will be sent to the server in waves up to the `batchLimit`.
      Defaults to 50.
      - parameter transaction: Treat as an all-or-nothing operation. If some operation failure occurs that
@@ -439,7 +443,7 @@ public extension Sequence where Element: ParseObject {
             if transaction {
                 batchLimit = commands.count
             } else {
-                batchLimit = limit != nil ? limit! : ParseConstants.batchLimit
+                batchLimit = limit ?? ParseConstants.batchLimit
             }
             let batches = BatchUtils.splitArray(commands, valuesPerSegment: batchLimit)
             var completed = 0
@@ -633,9 +637,12 @@ extension ParseObject {
     internal func ensureDeepSave(options: API.Options = [],
                                  completion: @escaping ([String: PointerType],
                                                         [UUID: ParseFile], ParseError?) -> Void) {
-
-        let queue = DispatchQueue(label: "com.parse.deepSave", qos: .default,
-                                  attributes: .concurrent, autoreleaseFrequency: .inherit, target: nil)
+        let uuid = UUID()
+        let queue = DispatchQueue(label: "com.parse.deepSave.\(uuid)",
+                                  qos: .default,
+                                  attributes: .concurrent,
+                                  autoreleaseFrequency: .inherit,
+                                  target: nil)
 
         queue.sync {
             var objectsFinishedSaving = [String: PointerType]()