+++ Sources/ParseSwift/Types/Query.swift
@@ -687,7 +687,7 @@ public struct Query<T>: Encodable, Equatable where T: ParseObject {
     }
 
     /**
-     Includes all nested `ParseObject`s.
+     Includes all nested `ParseObject`s one level deep.
      - warning: Requires Parse Server 3.0.0+
      */
     public func includeAll() -> Query<T> {
@@ -903,6 +903,77 @@ extension Query: Queryable {
         findCommand(explain: explain, hint: hint).executeAsync(options: options) { result in
             callbackQueue.async {
                 completion(result)
+            }
+        }
+    }
+
+    /**
+     Retrieves *asynchronously* a complete list of `ParseObject`'s  that satisfy this query.
+        
+      - parameter hint: String or Object of index that should be used when executing query.
+      - parameter batchLimit: The maximum number of objects to send in each batch. If the items to be batched.
+         is greater than the `batchLimit`, the objects will be sent to the server in waves up to the `batchLimit`.
+         Defaults to 50.
+      - parameter options: A set of header options sent to the server. Defaults to an empty set.
+      - parameter callbackQueue: The queue to return to after completion. Default value of .main.
+      - parameter completion: The block to execute.
+      It should have the following argument signature: `(Result<[Decodable], ParseError>)`.
+     - warning: The items are processed in an unspecified order. The query may not have any sort
+     order, and may not use limit or skip.
+    */
+    public func findAll(hint: String? = nil,
+                        batchLimit limit: Int? = nil,
+                        options: API.Options = [],
+                        callbackQueue: DispatchQueue = .main,
+                        completion: @escaping (Result<[ResultType], ParseError>) -> Void) {
+        if order != nil || skip > 0 || self.limit != 100 {
+            let error = ParseError(code: .unknownError,
+                             message: "Cannot iterate on a query with sort, skip, or limit.")
+            completion(.failure(error))
+            return
+        }
+        let uuid = UUID()
+        let queue = DispatchQueue(label: "com.parse.findAll.\(uuid)",
+                                  qos: .default,
+                                  attributes: .concurrent,
+                                  autoreleaseFrequency: .inherit,
+                                  target: nil)
+        queue.sync {
+
+            var query = self
+                .order([.ascending("objectId")])
+            query.limit = limit ?? ParseConstants.batchLimit
+            var results = [ResultType]()
+            var finished = false
+
+            while !finished {
+                do {
+                    let currentResults: [ResultType] = try query.findCommand(explain: false,
+                                                                             hint: hint).execute(options: options)
+                    results.append(contentsOf: currentResults)
+                    if currentResults.count >= query.limit {
+                        guard let lastObjectId = results[results.count - 1].objectId else {
+                            throw ParseError(code: .unknownError, message: "Last object should have an id.")
+                        }
+                        query.where.add("objectId" > lastObjectId)
+                    } else {
+                        finished = true
+                    }
+                } catch {
+                    callbackQueue.async {
+                        guard let parseError = error as? ParseError else {
+                            completion(.failure(ParseError(code: .unknownError,
+                                                           message: error.localizedDescription)))
+                            return
+                        }
+                        completion(.failure(parseError))
+                    }
+                    return
+                }
+            }
+
+            callbackQueue.async {
+                completion(.success(results))
             }
         }
     }